+++ agencies/routing.py
@@ -1,0 +1,9 @@
+"""WebSocket URL routing for the agencies application (council votes)."""
+
+from django.urls import re_path
+
+from . import consumers
+
+websocket_urlpatterns = [
+    re_path(r"ws/council/$", consumers.CouncilConsumer.as_asgi()),
+]